#33 Playlist download removed
* Lazy loading and download removed
* README.md update
* Code style fixes
