+++ src/main/java/com/github/kiulian/downloader/extractor/Extractor.java
@@ -31,7 +31,7 @@ public interface Extractor {
     void setRetryOnFailure(int retryOnFailure);
 
     String extractYtPlayerConfig(String html) throws YoutubeException;
-    
+
     String extractYtInitialData(String html) throws YoutubeException;
 
     String loadUrl(String url) throws YoutubeException;
