+++ src/main/java/com/github/kiulian/downloader/parser/Parser.java
@@ -1,7 +1,5 @@
 package com.github.kiulian.downloader.parser;
 
-import java.util.List;
-
 /*-
  * #
  * Java youtube video and audio downloader
@@ -29,19 +27,21 @@ import com.github.kiulian.downloader.extractor.Extractor;
 import com.github.kiulian.downloader.model.VideoDetails;
 import com.github.kiulian.downloader.model.formats.Format;
 import com.github.kiulian.downloader.model.playlist.PlaylistDetails;
-import com.github.kiulian.downloader.model.playlist.PlaylistVideo;
+import com.github.kiulian.downloader.model.playlist.PlaylistVideoDetails;
 import com.github.kiulian.downloader.model.subtitles.SubtitlesInfo;
+
+import java.util.List;
 
 public interface Parser {
 
     Extractor getExtractor();
 
     CipherFactory getCipherFactory();
-    
+
     /* Video */
 
     JSONObject getPlayerConfig(String htmlUrl) throws YoutubeException;
-    
+
     VideoDetails getVideoDetails(JSONObject config);
 
     String getJsUrl(JSONObject config) throws YoutubeException;
@@ -51,12 +51,12 @@ public interface Parser {
     List<SubtitlesInfo> getSubtitlesInfo(String videoId) throws YoutubeException;
 
     List<Format> parseFormats(JSONObject json) throws YoutubeException;
-    
+
     /* Playlist */
-    
+
     JSONObject getInitialData(String htmlUrl) throws YoutubeException;
-    
+
     PlaylistDetails getPlaylistDetails(String playlistId, JSONObject initialData);
-    
-    List<PlaylistVideo> getPlaylistVideos(JSONObject initialData, int size)  throws YoutubeException;
+
+    List<PlaylistVideoDetails> getPlaylistVideos(JSONObject initialData, int videoCount)  throws YoutubeException;
 }
